backend/requirements.txt
fastapi>=0.100.0
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
langgraph
langchain-openai
beautifulsoup4
requests
pydantic
pydantic-settings
websockets
python-dotenv
httpx
beautifulsoup4
pytest
pytest-asyncio
duckduckgo-search
ddgs
